Change data types from int back to long

+++ api/src/main/java/gov/va/vro/api/responses/ClaimMetricsResponse.java
@@ -17,8 +17,8 @@ import lombok.ToString;
 @ToString(includeFieldNames = true)
 public class ClaimMetricsResponse {
   private long claims;
-  private int totalEvidenceGenerations;
-  private int totalPdfGenerations;
+  private long totalEvidenceGenerations;
+  private long totalPdfGenerations;
 
   @Schema(description = "Error message in the case of an error")
   @JsonInclude(JsonInclude.Include.NON_EMPTY)
